Move computational model of decoder to separate class
Separate the logic of the decoder from the actual computational model.
In the future, this would allow to more easily change the model to something different, and also allow to apply Adaptive Computation Time on this model
