+++ seq2seq/models/__init__.py
@@ -1,4 +1,5 @@
 from .EncoderRNN import EncoderRNN
 from .DecoderRNN import DecoderRNN
+from .DecoderRNNModel import DecoderRNNModel
 from .TopKDecoder import TopKDecoder
 from .seq2seq import Seq2seq
